+++ AjoudongFE/app/src/main/AndroidManifest.xml
@@ -28,6 +28,13 @@
         <activity android:name=".ManagerMainActivity" />
         <activity android:name=".UserMainActivity" />
         <activity android:name=".SignupActivity"/>
+        <activity android:name=".AuthActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
