+++ AjoudongFE/app/src/main/java/com/example/ajoudongfe/SignupActivity.java
@@ -1,4 +1,14 @@
 package com.example.ajoudongfe;
 
-public class SignupActivity {
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class SignupActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_signup);
+    }
 }
