+++ AjoudongFE/app/src/main/java/com/example/ajoudongfe/UserMainActivity.java
@@ -80,7 +80,7 @@ public class UserMainActivity extends AppCompatActivity {
                 }
                 else if(id == R.id.user_logout){
                     Toast.makeText(context, "로그아웃중", Toast.LENGTH_SHORT).show();
-                    Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
+                    Intent intent = new Intent(getApplicationContext(), MainActivity.class);
                     intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP|Intent.FLAG_ACTIVITY_SINGLE_TOP);
                     startActivity(intent);
