--- AjoudongFE/app/src/main/res/layout/activity_login.xml
@@ -1,140 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <Button
-        android:id="@+id/loginButton"
-        android:layout_width="315dp"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="150dp"
-        android:background="@drawable/loginbutton"
-        android:text="로그인"
-        android:textColor="#ffffff"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.52"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/userLogin"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="13dp"
-        android:layout_marginEnd="13dp"
-        android:text="사용자 로그인"
-        app:layout_constraintEnd_toStartOf="@+id/managerLogin"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/findID" />
-
-    <Button
-        android:id="@+id/managerLogin"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="12dp"
-        android:layout_marginEnd="2dp"
-        android:text="간부 로그인"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/userLogin"
-        app:layout_constraintTop_toBottomOf="@+id/signup" />
-
-    <CheckBox
-        android:id="@+id/autoLogin"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="4dp"
-        android:layout_marginBottom="50dp"
-        android:buttonTint="#8EC0EB"
-        android:text="자동 로그인 하기"
-        app:layout_constraintBottom_toBottomOf="@+id/loginButton"
-        app:layout_constraintStart_toStartOf="@+id/loginButton" />
-
-    <ImageView
-        android:id="@+id/AJOUDONGLogo"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="31dp"
-        android:contentDescription="ajoudong logo"
-        app:layout_constraintBottom_toTopOf="@+id/idLayout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/logo_word" />
-
-    <TextView
-        android:id="@+id/signup"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="38dp"
-        android:layout_marginEnd="6dp"
-        android:text="회원가입"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/findPW"
-        app:layout_constraintTop_toBottomOf="@+id/loginButton" />
-
-    <TextView
-        android:id="@+id/findID"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:text="아이디 찾기"
-        app:layout_constraintBaseline_toBaselineOf="@+id/findPW"
-        app:layout_constraintEnd_toStartOf="@+id/findPW"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/findPW"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="38dp"
-        android:layout_marginEnd="19dp"
-        android:text="비밀번호 찾기"
-        app:layout_constraintEnd_toStartOf="@+id/signup"
-        app:layout_constraintStart_toEndOf="@+id/findID"
-        app:layout_constraintTop_toBottomOf="@+id/loginButton" />
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/idLayout"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="50dp"
-        android:layout_marginEnd="50dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.442">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/idInputText"
-            android:layout_width="match_parent"
-            android:layout_height="50dp"
-            android:background="@drawable/edittext_border"
-            android:hint="@string/IDTEXT"
-            android:inputType="textNoSuggestions" />/>
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/pwLayout"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="50dp"
-        android:layout_marginEnd="50dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/idLayout">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/pwInputText"
-            android:layout_width="match_parent"
-            android:layout_height="50dp"
-            android:background="@drawable/edittext_border"
-            android:hint="@string/PWTEXT"
-            android:inputType="textPassword|textNoSuggestions" />
-    </com.google.android.material.textfield.TextInputLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
